+++ src/components/shared/button/ButtonSpinner.tsx
@@ -1,7 +1,24 @@
-export function ButtonSpinner() {
+import { twMerge } from 'tailwind-merge'
+
+interface ButtonSpinnerProps {
+  bgStyle?: string
+  circleStyle?: string
+}
+
+export function ButtonSpinner({ bgStyle, circleStyle }: ButtonSpinnerProps) {
   return (
-    <div className="h-4 w-4 flex-initial animate-spin items-center justify-center rounded-full bg-gradient-to-tr from-transparent to-gray-300">
-      <div className="h-3 w-3 flex-initial rounded-full bg-gray-800"></div>
+    <div
+      className={twMerge(
+        'h-4 w-4 flex-initial animate-spin items-center justify-center rounded-full bg-gradient-to-tr from-transparent to-gray-300',
+        circleStyle
+      )}
+    >
+      <div
+        className={twMerge(
+          'h-3 w-3 flex-initial rounded-full bg-gray-800',
+          bgStyle
+        )}
+      ></div>
     </div>
   )
 }