+++ src/components/shared/sidebar/SidebarMenuItem.tsx
@@ -13,7 +13,13 @@ import { MENU_LINKS, SOON_ITEMS } from '@/constants'
 
 import { useAppSelector } from '@/store'
 
-import { Badge, Link, ModalSupport } from '@/components/shared'
+import {
+  Badge,
+  DrawerNotifications,
+  Link,
+  ModalLogout,
+  ModalSupport,
+} from '@/components/shared'
 
 import { SidebarMenuItemIcon } from './SidebarMenuItemIcon'
 
@@ -36,6 +42,8 @@ export function SidebarMenuItem({ item }: SidebarMenuItemProps) {
   const notifications = useAppSelector((state) => state.notifications)
 
   const [openModalSupport, setOpenModalSupport] = useState(false)
+  const [openModalLogout, setOpenModalLogout] = useState(false)
+  const [openDrawerNotifications, setOpenDrawerNotifications] = useState(false)
 
   const isSoon = SOON_ITEMS.includes(item)
 
@@ -56,7 +64,7 @@ export function SidebarMenuItem({ item }: SidebarMenuItemProps) {
         break
 
       case 'notificações':
-        console.log(item)
+        setOpenDrawerNotifications(true)
         break
 
       case 'suporte':
@@ -64,7 +72,7 @@ export function SidebarMenuItem({ item }: SidebarMenuItemProps) {
         break
 
       case 'sair':
-        console.log(item)
+        setOpenModalLogout(true)
         break
 
       default:
@@ -145,6 +153,12 @@ export function SidebarMenuItem({ item }: SidebarMenuItemProps) {
       </Link>
 
       <ModalSupport open={openModalSupport} setOpen={setOpenModalSupport} />
+      <ModalLogout open={openModalLogout} setOpen={setOpenModalLogout} />
+      <ModalLogout open={openModalLogout} setOpen={setOpenModalLogout} />
+      <DrawerNotifications
+        open={openDrawerNotifications}
+        setOpen={setOpenDrawerNotifications}
+      />
     </div>
   )
 }