+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Cria componente de `ScrollArea` no projeto next.
+- Cria componente de `DrawerNotification` e todos seus componentes no projeto next [#728](https://github.com/3C-gg/reload-frontend/issues/729).
+- Cria componente de `Drawer` no projeto next.
+- Cria componente de `ModalLogout` no projeto next [#728](https://github.com/3C-gg/reload-frontend/issues/728).
 - Cria componente de `ModalSupport` e todos seus componentes no projeto next [#731](https://github.com/3C-gg/reload-frontend/issues/731).
 - Cria componente de `sidebar` no projeto next [#713](https://github.com/3C-gg/reload-frontend/issues/713).
 - Cria componente de `Timer` no projeto next.
@@ -46,6 +50,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Adiciona variant `neutral` em ` Button`.
 - Ajusta `imports` usando alias no projeto next.
 - Ajusta `httpService` no projeto next.
 - Altera tipagem de `httpService` no projeto next.
@@ -97,6 +102,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- Corrige slice de `notifications` no projeto next.
 - Corrige bordas do componente `Progress`.
 - Corrige componente `Progress` para não pintar de vermelho quando os `pontos ganhos (earned_points)` forem 0 [#686](https://github.com/3C-gg/reload-frontend/issues/686).
 - Corrige `layout` dos `stats` da esquerda na pagina de `/profile` que estava esticando junto com o histórico de partidas [#689](https://github.com/3C-gg/reload-frontend/issues/689).